+++ src/framex/plugin/__init__.py
@@ -1,12 +1,10 @@
 from contextvars import ContextVar
-from typing import Any, Optional, cast
+from typing import Optional
 
-from aiocache import Cache, cached
-from ray import serve
 from ray.serve.handle import DeploymentHandle
 
-from framex.config import settings
-from framex.consts import APP_NAME, PROXY_PLUGIN_NAME
+from framex.adapter import get_adapter
+from framex.consts import PROXY_PLUGIN_NAME
 from framex.log import logger
 from framex.plugin.manage import PluginManager
 from framex.plugin.model import Plugin, PluginApi
@@ -20,7 +18,8 @@ def get_loaded_plugins() -> set["Plugin"]:
     return set(_manager._plugins.values())
 
 
-def init_all_deployments() -> list[DeploymentHandle]:
+@logger.catch()
+def init_all_deployments(enable_proxy: bool) -> list[DeploymentHandle]:
     deployments = []
     for plugin in get_loaded_plugins():
         for dep in plugin.deployments:
@@ -31,20 +30,25 @@ def init_all_deployments() -> list[DeploymentHandle]:
                 if api_name in remote_apis:
                     continue
 
-                if (not api_name.startswith("/")) and not settings.enable_proxy:
-                    raise RuntimeError(f"Required remote api({api_name}) not found")
+                if api_name.startswith("/") and enable_proxy:
+                    remote_apis[api_name] = PluginApi(
+                        api=api_name,
+                        deployment_name=PROXY_PLUGIN_NAME,
+                        call_type=ApiType.PROXY,
+                        plugin_name=PROXY_PLUGIN_NAME,
+                    )
+                    logger.opt(colors=True).warning(
+                        f"Api(<r>{api_name}</r>) not found, "
+                        f"plugin(<r>{dep.deployment}</r>) will "
+                        f"use proxy plugin({PROXY_PLUGIN_NAME}) to transfer!"
+                    )
+                else:
+                    raise RuntimeError(
+                        f"Plugin({dep.deployment}) init failed, Required remote api({api_name}) not found"
+                    )
+            deployment = get_adapter().bind(dep.deployment, remote_apis=remote_apis, config=plugin.config)
 
-                remote_apis[api_name] = PluginApi(
-                    api=api_name,
-                    deployment_name=PROXY_PLUGIN_NAME,
-                    call_type=ApiType.PROXY,
-                    plugin_name=PROXY_PLUGIN_NAME,
-                )
-                logger.opt(colors=True).warning(
-                    f"Api(<y>{api_name}</y>) not found, "
-                    f"plugin(<y>{dep.deployment.name}</y>) will use proxy plugin({PROXY_PLUGIN_NAME}) to transfer!"
-                )
-            deployments.append(dep.deployment.bind(remote_apis=remote_apis, config=plugin.config))
+            deployments.append(deployment)
 
     return deployments
 
@@ -53,42 +57,14 @@ def get_http_plugin_apis() -> list["PluginApi"]:
     return _manager.http_plugin_apis
 
 
-@cached(cache=Cache.MEMORY)
-async def _check_is_gen_api(path: str) -> bool:
-    handle: DeploymentHandle = serve.get_deployment_handle(PROXY_PLUGIN_NAME, app_name=APP_NAME)
-    c_handle = handle.check_is_gen_api
-    return cast(bool, await c_handle.remote(path=path))
-
-
-async def call_remote_api(api: PluginApi, **kwargs: Any) -> Any:
-    handle: DeploymentHandle = serve.get_deployment_handle(api.deployment_name, app_name=APP_NAME)
-    c_handle = getattr(handle, api.func_name)
-    if not c_handle:
-        raise RuntimeError(f"No handle found for func_name({api.func_name})")
-
-    stream = api.stream
-
-    if api.call_type == ApiType.PROXY:
-        kwargs["proxy_path"] = api.api
-        stream = await _check_is_gen_api(api.api)
-
-    if stream:
-        gen = c_handle.options(stream=True).remote(**kwargs)
-        return [chunk async for chunk in gen]
-
-    res = await c_handle.remote(**kwargs)
-    return data if api.call_type == ApiType.PROXY and (data := res.get("data")) else res
-
-
 from .base import BasePlugin
 from .load import load_builtin_plugin, load_plugins
-from .model import ApiType, PluginApi, PluginMetadata
+from .model import ApiType, PluginMetadata
 from .on import on_register, on_request
 
 __all__ = [
     "ApiType",
     "BasePlugin",
-    "PluginApi",
     "PluginMetadata",
     "load_builtin_plugin",
     "load_plugins",