+++ src/framex/plugin/on.py
@@ -3,8 +3,8 @@ from collections.abc import Callable
 from typing import Any, get_type_hints
 
 from pydantic import BaseModel
-from ray import serve
 
+from framex.adapter import get_adapter
 from framex.consts import API_STR
 from framex.plugin.model import ApiType, PluginApi, PluginDeployment
 from framex.utils import extract_method_params, plugin_to_deployment_name
@@ -47,13 +47,10 @@ def on_register(**kwargs: Any) -> Callable[[type], type]:
                         )
                     )
 
-            cls = serve.deployment(**kwargs)(cls)
+            cls = get_adapter().to_deployment(cls, **kwargs)
+
             deployment = PluginDeployment(plugin_apis=plugin_apis, deployment=cls)
             plugin.deployments.append(deployment)
-
-            # logger.opt(colors=True).debug(
-            #     f'Found deploment "<m>{escape_tag(kwargs["name"])}</m> " from {plugin.module_name}'
-            # )
 
         return cls
 
@@ -61,7 +58,7 @@ def on_register(**kwargs: Any) -> Callable[[type], type]:
 
 
 def on_request(
-    path: str | None = None, methods: list[str] | None = None, call_type: ApiType = ApiType.ALL, stream: bool = False
+    path: str | None = None, methods: list[str] | None = None, call_type: ApiType = ApiType.HTTP, stream: bool = False
 ) -> Callable:
     if methods is None:
         methods = ["GET"]