+++ src/framex/plugins/echo.py
@@ -2,8 +2,11 @@ import asyncio
 from collections.abc import AsyncGenerator
 from typing import Any
 
+from pydantic import BaseModel
+
 from framex.consts import VERSION
 from framex.plugin import BasePlugin, PluginMetadata, on_register, on_request
+from framex.plugin.model import ApiType
 from framex.utils import StreamEnventType, make_stream_event
 
 __plugin_meta__ = PluginMetadata(
@@ -16,6 +19,11 @@ __plugin_meta__ = PluginMetadata(
 )
 
 
+class EchoModel(BaseModel):
+    id: int
+    name: str = "原神"
+
+
 @on_register()
 class EchoPlugin(BasePlugin):
     def __init__(self, **kwargs: Any) -> None:
@@ -25,9 +33,17 @@ class EchoPlugin(BasePlugin):
     async def echo(self, message: str) -> str:
         return message
 
+    @on_request("/echo_model", methods=["POST"])
+    async def echo_model(self, message: str, model: EchoModel) -> str:
+        return f"{message},{model.model_dump()}"
+
     @on_request("/echo_stream", methods=["GET"], stream=True)
     async def echo_stream(self, message: str) -> AsyncGenerator[str, None]:
         for char in f"原神真好玩呀, {message}":
             yield make_stream_event(StreamEnventType.MESSAGE_CHUNK, char)
             await asyncio.sleep(0.1)
-        yield make_stream_event(StreamEnventType.FINISH, char)
+        yield make_stream_event(StreamEnventType.FINISH)
+
+    @on_request(call_type=ApiType.FUNC)
+    async def confess(self, message: str) -> str:
+        return f"我是原神哟! 收到您的消息{message}"