+++ src/framex/plugins/proxy/__init__.py
@@ -6,9 +6,10 @@ from typing import Any, cast
 import httpx
 from pydantic import BaseModel, create_model
 
-from framex.consts import APP_NAME, BACKEND_NAME, PROXY_PLUGIN_NAME, VERSION
+from framex.adapter import get_adapter
+from framex.consts import BACKEND_NAME, PROXY_PLUGIN_NAME, VERSION
 from framex.log import logger
-from framex.plugin import BasePlugin, PluginApi, PluginMetadata, call_remote_api, on_register
+from framex.plugin import BasePlugin, PluginApi, PluginMetadata, on_register
 from framex.plugin.model import ApiType
 from framex.plugin.on import on_request
 from framex.plugins.proxy.builder import create_pydantic_model, type_map
@@ -37,6 +38,7 @@ class ProxyPlugin(BasePlugin):
         super().__init__(**kwargs)
 
     async def on_start(self) -> None:
+        await super().on_start()
         if not self.config.proxy_urls:
             logger.warning("No url provided, skipping proxy plugin")
             return
@@ -54,6 +56,8 @@ class ProxyPlugin(BasePlugin):
         return cast(dict[str, Any], response.json())
 
     async def _parse_openai_docs(self, url: str) -> None:
+        adapter = get_adapter()
+
         openapi_data = await self._get_openai_docs(url)
         paths = openapi_data.get("paths", {})
         components = openapi_data.get("components", {}).get("schemas", {})
@@ -83,7 +87,7 @@ class ProxyPlugin(BasePlugin):
                     Model = create_pydantic_model(schema_name, model_schema, components)  # noqa
                     params.append(("model", Model))
 
-                logger.opt(colors=True).info(f"Found proxy api({method}) <y>{url}{path}</y>")
+                logger.opt(colors=True).debug(f"Found proxy api({method}) <y>{url}{path}</y>")
 
                 func_name = body.get("operationId")
                 is_stream = path in self.config.force_stream_apis
@@ -95,10 +99,9 @@ class ProxyPlugin(BasePlugin):
                     deployment_name=BACKEND_NAME,
                     func_name="register_route",
                 )
-                from ray import serve
 
-                handle = serve.get_deployment_handle(PROXY_PLUGIN_NAME, app_name=APP_NAME)
-                await call_remote_api(
+                handle = adapter.get_handle(PROXY_PLUGIN_NAME)
+                await adapter.call_func(
                     plugin_api,
                     path=path,
                     methods=[method],